refactor(Pagination): rename to ts
