+++ components/pagination/style.ts
