+++ components/pagination/types.ts
